+++ PoGo.NecroBot.CLI/Program.cs
@@ -87,15 +87,13 @@ namespace PoGo.NecroBot.CLI
                         break;
                 }
             }
+            
+            Logger.AddLogger(new ConsoleLogger(LogLevel.Service), _subPath);
+            Logger.AddLogger(new FileLogger(LogLevel.Service), _subPath);
+            Logger.AddLogger(new WebSocketLogger(LogLevel.Service), _subPath);
 
-
-
-            Logger.SetLogger(new ConsoleLogger(LogLevel.Service), _subPath);
-
-            // We can now rely on killswitch from checking the minimum client version returned by the API.
-            // A manually triggered killswitch is no longer needed.
-            //if (!_ignoreKillSwitch && CheckKillSwitch() || CheckMKillSwitch())
-                //return;
+            if (!_ignoreKillSwitch && CheckKillSwitch() || CheckMKillSwitch())
+                return;
 
             var profilePath = Path.Combine(Directory.GetCurrentDirectory(), _subPath);
             var profileConfigPath = Path.Combine(profilePath, "config");
@@ -379,10 +377,6 @@ namespace PoGo.NecroBot.CLI
 
         private static bool CheckKillSwitch()
         {
-            #if DEBUG
-                return false;
-            #endif
-
             using (var wC = new WebClient())
             {
                 try