+++ PoGo.NecroBot.Logic/Tasks/TransferDuplicatePokemonTask.cs
@@ -18,6 +18,8 @@ namespace PoGo.NecroBot.Logic.Tasks
         {
             cancellationToken.ThrowIfCancellationRequested();
 
+            if (!session.LogicSettings.TransferDuplicatePokemon) return;
+
             if (session.LogicSettings.AutoFavoritePokemon)
                 await FavoritePokemonTask.Execute(session, cancellationToken);